+++ src/main/java/apple/excursion/discord/commands/general/benchmark/CommandCalendar.java
@@ -1,7 +1,7 @@
-package apple.excursion.discord.commands.general.leaderboard;
+package apple.excursion.discord.commands.general.benchmark;
 
 import apple.excursion.discord.commands.DoCommand;
-import apple.excursion.discord.reactions.messages.leaderboard.CalendarMessage;
+import apple.excursion.discord.reactions.messages.benchmark.CalendarMessage;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 
 public class CommandCalendar implements DoCommand {
