+++ src/main/java/apple/excursion/discord/commands/general/benchmark/CommandGuildLeaderboard.java
@@ -1,4 +1,4 @@
-package apple.excursion.discord.commands.general.leaderboard;
+package apple.excursion.discord.commands.general.benchmark;
 
 import apple.excursion.database.objects.guild.GuildHeader;
 import apple.excursion.database.queries.GetDB;
@@ -7,8 +7,8 @@ import apple.excursion.database.objects.player.PlayerData;
 import apple.excursion.discord.commands.DoCommand;
 import apple.excursion.database.objects.guild.GuildLeaderboardEntry;
 import apple.excursion.database.objects.guild.LeaderboardOfGuilds;
-import apple.excursion.discord.reactions.messages.leaderboard.GuildLeaderboardMessage;
-import apple.excursion.discord.reactions.messages.leaderboard.GuildProfileMessage;
+import apple.excursion.discord.reactions.messages.benchmark.GuildLeaderboardMessage;
+import apple.excursion.discord.reactions.messages.benchmark.GuildProfileMessage;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 
 import java.sql.SQLException;
