+++ src/main/java/apple/excursion/discord/commands/general/benchmark/CommandHistory.java
@@ -1,9 +1,8 @@
-package apple.excursion.discord.commands.general.leaderboard;
+package apple.excursion.discord.commands.general.benchmark;
 
 import apple.excursion.discord.commands.DetermineArguments;
 import apple.excursion.discord.commands.DoCommand;
-import apple.excursion.discord.reactions.messages.history.GuildHistoryMessage;
-import apple.excursion.discord.reactions.messages.history.HistoryMessage;
+import apple.excursion.discord.reactions.messages.benchmark.history.HistoryMessage;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 
 import java.util.ArrayList;
