+++ src/main/java/apple/excursion/discord/reactions/messages/benchmark/CalendarMessage.java
@@ -1,4 +1,4 @@
-package apple.excursion.discord.reactions.messages.leaderboard;
+package apple.excursion.discord.reactions.messages.benchmark;
 
 import apple.excursion.database.queries.GetCalendarDB;
 import apple.excursion.discord.data.Task;
