+++ src/main/java/apple/excursion/discord/reactions/messages/benchmark/ProfileMessage.java
@@ -1,4 +1,4 @@
-package apple.excursion.discord.reactions.messages.self;
+package apple.excursion.discord.reactions.messages.benchmark;
 
 import apple.excursion.database.objects.OldSubmission;
 import apple.excursion.database.objects.player.PlayerData;
