+++ src/main/java/apple/excursion/discord/reactions/messages/benchmark/history/GuildHistoryMessage.java
@@ -1,10 +1,10 @@
-package apple.excursion.discord.reactions.messages.history;
+package apple.excursion.discord.reactions.messages.benchmark.history;
 
 import apple.excursion.database.queries.GetDB;
-import apple.excursion.discord.data.answers.HistoryPlayerLeaderboard;
+import apple.excursion.discord.data.answers.HistoryLeaderboardOfGuilds;
 import apple.excursion.discord.reactions.AllReactables;
 import apple.excursion.discord.reactions.ReactableMessage;
-import apple.excursion.discord.reactions.messages.leaderboard.LeaderboardMessage;
+import apple.excursion.discord.reactions.messages.benchmark.GuildLeaderboardMessage;
 import apple.excursion.utils.Pretty;
 import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.entities.MessageChannel;
@@ -16,12 +16,13 @@ import java.util.Calendar;
 import java.util.HashMap;
 import java.util.Map;
 
-import static apple.excursion.discord.reactions.messages.leaderboard.CalendarMessage.EPOCH_START_OF_EXCURSION;
-import static apple.excursion.discord.reactions.messages.leaderboard.CalendarMessage.EPOCH_START_OF_SUBMISSION_HISTORY;
+import static apple.excursion.discord.reactions.messages.benchmark.CalendarMessage.EPOCH_START_OF_EXCURSION;
+import static apple.excursion.discord.reactions.messages.benchmark.CalendarMessage.EPOCH_START_OF_SUBMISSION_HISTORY;
 
-public class HistoryMessage implements ReactableMessage {
+
+public class GuildHistoryMessage implements ReactableMessage {
     private final Message message;
-    private final Map<Long, HistoryPlayerLeaderboard> leaderboard = new HashMap<>();
+    private final Map<Long, HistoryLeaderboardOfGuilds> leaderboard = new HashMap<>();
     private final Calendar timeLookingAt = Calendar.getInstance();
     private final Calendar later;
     private final int timeInterval;
@@ -29,7 +30,7 @@ public class HistoryMessage implements ReactableMessage {
     private long lastUpdated = System.currentTimeMillis();
     private int page = 0;
 
-    public HistoryMessage(MessageChannel channel, int timeInterval, int timeField) {
+    public GuildHistoryMessage(MessageChannel channel, int timeInterval, int timeField) {
         later = Calendar.getInstance();
         later.setTimeInMillis(timeLookingAt.getTimeInMillis());
         later.add(timeField, timeInterval);
@@ -50,10 +51,10 @@ public class HistoryMessage implements ReactableMessage {
         if (later.getTimeInMillis() <= timeLookingAt.getTimeInMillis()) {
             return "```\nFuture\nThis data for this time is not available yet\n```";
         }
-        HistoryPlayerLeaderboard myLeaderboard = leaderboard.get(timeLookingAt.getTimeInMillis());
+        HistoryLeaderboardOfGuilds myLeaderboard = leaderboard.get(timeLookingAt.getTimeInMillis());
         if (myLeaderboard == null) {
             try {
-                myLeaderboard = GetDB.getPlayerLeaderboard(timeField, timeInterval, timeLookingAt);
+                myLeaderboard = GetDB.getGuildLeaderboard(timeField, timeInterval, timeLookingAt);
                 leaderboard.put(timeLookingAt.getTimeInMillis(), myLeaderboard);
 
             } catch (SQLException throwables) {
@@ -69,7 +70,7 @@ public class HistoryMessage implements ReactableMessage {
                 page + 1,
                 Pretty.date(startTime),
                 Pretty.date(endTime));
-        return LeaderboardMessage.makeMessageStatic(myLeaderboard.leaderboard, page, title);
+        return GuildLeaderboardMessage.makeMessageStatic(myLeaderboard.leaderboard, page, title);
     }
 
     @Override
@@ -121,8 +122,8 @@ public class HistoryMessage implements ReactableMessage {
     }
 
     private void right() {
-        HistoryPlayerLeaderboard myLeaderboard = leaderboard.get(timeLookingAt.getTimeInMillis());
-        if ((myLeaderboard.leaderboard.size() - 1) / LeaderboardMessage.ENTRIES_PER_PAGE >= page + 1) {
+        HistoryLeaderboardOfGuilds myLeaderboard = leaderboard.get(timeLookingAt.getTimeInMillis());
+        if ((myLeaderboard.leaderboard.size() - 1) / GuildLeaderboardMessage.ENTRIES_PER_PAGE >= page + 1) {
             page++;
             message.editMessage(makeMessage()).queue();
         }
