+++ src/main/java/apple/excursion/discord/reactions/messages/postcard/SubmissionMessage.java
@@ -1,4 +1,4 @@
-package apple.excursion.discord.reactions.messages.settings;
+package apple.excursion.discord.reactions.messages.postcard;
 
 import apple.excursion.database.queries.InsertDB;
 import apple.excursion.discord.data.answers.SubmissionData;
@@ -13,8 +13,6 @@ import net.dv8tion.jda.api.events.message.react.MessageReactionAddEvent;
 
 import java.sql.SQLException;
 import java.util.stream.Collectors;
-
-import static apple.excursion.discord.commands.general.settings.CommandSubmit.BOT_COLOR;
 
 public class SubmissionMessage implements ReactableMessage {
     private SubmissionData data;
